07_Modulo_7/03_SEMANA_03/my_project/app/requirements.txt
fastapi
uvicorn
python-jose[cryptographjy]
passlib[bcrypt]